nao sei o que escrever

+++ main.c
@@ -5,8 +5,8 @@
 #define MAXIMODOSDADOS 2000
 
 // Calculando a média aritmética
-float calcular_media(const float *dados, int n) {
-    float soma = 0.0f;
+double calcular_media(const double *dados, int n) {
+    double soma = 0.0;
     for (int i = 0; i < n; i++) {
         soma += dados[i];
     }
@@ -14,23 +14,23 @@ float calcular_media(const float *dados, int n) {
 }
 
 // Calculando o desvio padrão
-float calcular_desvio_padrao(const float *dados, int n, float media) {
-    float soma_diferenca_quadrado = 0.0f;
+double calcular_desvio_padrao(const double *dados, int n, double media) {
+    double soma_diferenca_quadrado = 0.0;
     for (int i = 0; i < n; i++) {
-        float diff = dados[i] - media;
+        double diff = dados[i] - media;
         soma_diferenca_quadrado += diff * diff;
     }
-    return sqrtf(soma_diferenca_quadrado / n);
+    return sqrt(soma_diferenca_quadrado / n);
 }
 
 // Faz a normalização z score
-void normalizar_zscore(const float *entrada, float *saida, int n) {
-    float media = calcular_media(entrada, n);
-    float desvio = calcular_desvio_padrao(entrada, n, media);
+void normalizar_zscore(const double *entrada, double *saida, int n) {
+    double media = calcular_media(entrada, n);
+    double desvio = calcular_desvio_padrao(entrada, n, media);
 
-    if (desvio == 0.0f) {
+    if (desvio == 0.0) {
         for (int i = 0; i < n; i++) {
-            saida[i] = 0.0f;
+            saida[i] = 0.0;
         }
     } else {
         for (int i = 0; i < n; i++) {
@@ -40,12 +40,12 @@ void normalizar_zscore(const float *entrada, float *saida, int n) {
 }
 
 int main() {
-    float dados[MAXIMODOSDADOS];
-    float normalizado[MAXIMODOSDADOS]; // vetor de saída
+    double dados[MAXIMODOSDADOS];
+    double normalizado[MAXIMODOSDADOS]; // vetor de saída
     FILE *arquivo_entrada;
     FILE *arquivo_saida;
     int i = 0;
-    float valor_temporario; // para armazenar o valor lido
+    double valor_temporario; // para armazenar o valor lido
     int valores_lidos = 0;
 
     // Abre o arquivo para leitura
@@ -58,7 +58,7 @@ int main() {
     }
 
     // Ler o CSV
-    while (valores_lidos < MAXIMODOSDADOS && fscanf(arquivo_entrada, "%f", &valor_temporario) == 1) {
+    while (valores_lidos < MAXIMODOSDADOS && fscanf(arquivo_entrada, "%lf", &valor_temporario) == 1) {
         dados[valores_lidos] = valor_temporario;
         valores_lidos++;
         int c;
@@ -83,7 +83,7 @@ int main() {
 
     // Escrevendo os valores normalizados no arquivo CSV de saida
     for (i = 0; i < valores_lidos; i++) {
-        fprintf(arquivo_saida, "%.6f", normalizado[i]);
+        fprintf(arquivo_saida, "%lf", normalizado[i]);
         if (i < valores_lidos - 1) fprintf(arquivo_saida, ",");
     }
     fprintf(arquivo_saida, "\n"); //Quebra de linha no final
